news_stream_streams/requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.10
confluent-kafka==1.8.2
idna==3.3
iniconfig==1.1.1
mongoengine==0.23.1
oauthlib==3.1.1
packaging==21.3
pluggy==1.0.0
prometheus-client==0.12.0
py==1.11.0
pymongo==3.12.3
pyparsing==3.0.7
pytest==6.2.5
python-dateutil==2.8.2
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
toml==0.10.2
tweepy==4.4.0
urllib3==1.26.8